app/utils/reader/requirements.txt
pika
mysql-connector-python
python-dotenv
nfcpy